feat(client)!: remove awaitable functionality from WebSocket connections
This commit removes the `Awaitable` wrapper from the WebSocket connection API, simplifying the interface and making it more explicit. Previously, the `connect` function would return an awaitable object that would automatically open the connection when awaited. Now, users must explicitly call the `open` method on the returned WebSocketChannel instance.

The change affects both the `Client.connect` method and the standalone `connect` function. This is a breaking change as code that relied on awaiting the connection object will need to be updated to explicitly call `open()`.