chore: update package.json files
The root `package.json` file was updated to remove the `DEBUG` environment variable from the `lint` and `lint:fix` scripts. This change ensures that the linting process runs without unnecessary debug information, which can clutter the output and slow down the linting process.

Additionally, the `package.json` file for the `fs` package was updated to include `memfs` as a peer dependency. This change ensures that `memfs` is correctly managed as a peer dependency, which is important for maintaining compatibility and avoiding conflicts with other packages that may also depend on `memfs`. This update helps to improve the overall stability and maintainability of the `fs` package.